FIX #99 - the properties `maxPingOut` and `pingInterval` shadow the properties (#100)
for plain nats client pings. To enable them to be configured separately, they
are renamed to `stanMaxPingOut` and `stanPingInterval`.

FIX #98 - node-nats options are not exposed in the index.d.ts, so typescript
aware editors/ides show errors on unknown properties.

FIX #97 - invalidated by renaming the options

+++ test/connect.js
@@ -47,8 +47,11 @@ describe('Basic Connectivity', function() {
   });
 
   // Shutdown our server after we are done
-  after(function(){
-    server.kill();
+  after(function(done){
+    if(server) {
+      server.kill();
+    }
+    done();
   });
